+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Add grid placement mode.
+- Add angle locking.
 
 ### 0.9.8.6
 - Update for game version 1.2.0.